Change inserting data method
A new row is inserted at the top of the sheet and data written there, instead of finding the next free row and writing data there.

This speeds up the process and makes the sheet more user friendly when large amounts of data are present.  This method seems to be good for up to around 1000 rows and then issues with time outs may start occurring.  Code to add data to the next free row is still present, just commented out